-User.java
-AdminUser, EditorUser, GuestUser estendendo User (sem duplicar estado).

-Testes JUnit passando (vide UserStructureTest).
